refactor: rename graupel libs -> icon_graupel
