+++ tests/test_graupel_microphysics_scheme.py
@@ -23,7 +23,7 @@ import os
 
 path = os.environ.get("PY_GRAUPEL_DIR")
 if path and path is not None:
-    from libs.graupel.microphysics_scheme_wrapper import (
+    from libs.icon_graupel.microphysics_scheme_wrapper import (
         MicrophysicsSchemeWrapper,
         py_graupel,
     )
